core: Implement caching
Implements caching based on task objects. Each task object is assigned
a hash (based on the sha1 algorithm in combination with pickle), that
is stored inside a dict. Each hash maps to a list of results previously
returned by a bear.

To make caching work properly with the existing bear base types
(FileBear, ProjectBear, DependencyBear), the task optimizations
with custom `execute_task()` function were removed. The keyword
arguments have to be passed with the tasks, otherwise those settings
don't influence the cache-hash and changes in them would cause a
wrong cache hit.

Closes https://github.com/coala/coala/issues/4350