Add NlSection and Parsers Functionality
Add NlSection, NlSectionPosition, Parser, PyJinjaParser
Add regex module into requirement.txt

NlSection: This is used to maintain the position of
nested language section in the original nested language
file.This encompasses of start and end indices which
keep track of the section's start and end.

NlSectionPosition: This stores the start and end position
of a particular section.

Parser: This is the super class for all the parsers.

PyJinjaParser: This parser parses the file that contains both
Python and Jinja. This parses the original file and spits out
NlSections.