NlCore: Add NlCore
This PR adds NlCore

NlCore is the main core of the entire Nested language
Architecture. This is the class with with the coala
interacts with, to access an functions of Nested language.

This is responsible for creating coala_nl_sections and
writing the selected patches by the user to the original
file.