Parser.py: Add Parser.py
This PR adds the Parser class.

Parsers are used to seperate a nested language file into
various nested language sections. Parser eats up the files
and spits out a list of NlSection objects that is used to
identify which part of the file has which language.

This is a Base Class for all the parsers that would henceforth
parse nested language files.