ConsoleInteraction: Add support for Nested langs
This PR adds the support for nested languages mode in
ConsoleInteraction.py. The function have been updated
to pass the nl_file_dict where the patches chosen by
the user will be applied.

We have also disabled printing the affected files in
Nested Language Mode, since we do not physically have
these files in the system.