Processing.py: Add Support for Nested lang mode
This PR adds the support for nested language mode. It
allows creation of new nl_sections when the Nested language
mode is run.

During the execution of instantiate_processes, we get the
file_dict of the temporary files from the nested lang parsers
instead of fetching it from the physical file.